requirements.txt
Django==3.0.7
python-dotenv==0.13.0
pytz==2019.3
django-crispy-forms==1.9.1
Pillow==7.1.2
python-dotenv==0.13.0
